Throughput: bf16 autocast + TF32 behind --precision; record batch_size/precision
The driver ran fp32 with no autocast, no TF32 and no compile, and every forward at
batch dim 1: 16 teacher + 160 thought-loop + 16 final = 192 sequential batch-1
forwards per optimizer step (~960k per arm-seed over 5,000 steps) at ~9.6e13
FLOPs/step. On an A100 that is 309 ms/step at bf16 peak vs 4.9 s at strict
fp32-no-TF32 peak, so most of the tensor-core throughput was being left unused.

- train_driver.py: configure_precision() (TF32) + autocast_ctx() (bf16 autocast,
  CUDA-only so CPU numerics are untouched), applied to the training forward and to
  in-loop val scoring. backward() stays outside autocast. No GradScaler needed for
  bf16. Both live in the shared driver, so precision is arm-invariant by
  construction.
- loss.py: pin the distill smooth-L1 and R1's 100k-way softmax to fp32 -- bf16
  would quantize exactly the small teacher/student differences distill exists to
  close. The .float() calls are no-ops on the fp32 path, which stays bit-identical.
- train_codi.py: --precision {fp32,bf16} (default bf16), also wrapping the final
  gate scoring so it matches how best.pt was selected. Records precision and
  batch_size in metrics.json (batch_size was previously unrecorded).

Also documents, with measurements, that --batch-size is NOT a throughput knob:
examples are processed one at a time, so per-example step time is flat from batch 2
upward (155.6 -> 71.1 / 71.7 / 64.0 ms at 1/2/4/8) while total step time grows
linearly. At 16 the fixed per-step optimizer cost is already amortized.

Not done, deliberately: sequence packing via cu_doc_lens (public on
Attention.forward, no core change needed) and the 9.7x thought-loop recompute.
Core's KVCacheManager cannot fix the latter -- it writes in-place into registered
buffers, i.e. it is an inference cache, and CODI training needs gradients through
all K steps. It is also unverifiable here for eval use: there is no
Transformer-level API and TorchAttentionBackend raises "doesn't support KV
caching", so the CPU path cannot exercise it. Deferred rather than shipped
untested on the code path that produces the gates.

140 latentcot tests pass; isort + black clean.

+++ src/olmo_core/latentcot/loss.py
@@ -63,18 +63,20 @@ def vocab_manifold_reg(
     if kind == "none":
         return thoughts.new_zeros(())
     if kind == "L2":
-        return thoughts.pow(2).mean()
+        return thoughts.float().pow(2).mean()
 
     logits = model.lm_head(thoughts)  # (batch, K, vocab) — labels=None returns logits
     embeddings = model.embeddings.weight  # (vocab, d_model)
     if kind == "R2":
         target = embeddings[logits.argmax(dim=-1)]  # (batch, K, d_model)
-        return (thoughts - target).pow(2).mean()
+        return (thoughts.float() - target.float()).pow(2).mean()
 
-    # R1
-    probs = torch.softmax(logits, dim=-1)
-    target = probs @ embeddings  # (batch, K, d_model)
-    reg = (thoughts - target).pow(2).mean()
+    # R1. The softmax runs in fp32: under bf16 autocast a 100k-way softmax loses enough
+    # precision in the tail to move the mixture target, and the .float() calls are no-ops
+    # in the fp32 path (so both paths stay bit-identical to before).
+    probs = torch.softmax(logits.float(), dim=-1)
+    target = probs @ embeddings.float()  # (batch, K, d_model)
+    reg = (thoughts.float() - target).pow(2).mean()
     if entropy_floor > 0:
         entropy = -(probs * probs.clamp_min(1e-9).log()).sum(dim=-1).mean()
         reg = reg + torch.relu(torch.as_tensor(entropy_floor, device=reg.device) - entropy)
@@ -186,7 +188,9 @@ def codi_loss(
         )
 
         # --- Distillation + vocab regularizer ---
-        distill = F.smooth_l1_loss(student_acts, teacher_acts.detach())
+        # fp32: this aligns raw hidden states across all layers, and bf16's ~3 decimal digits
+        # would quantize exactly the small teacher-student differences the term exists to close.
+        distill = F.smooth_l1_loss(student_acts.float(), teacher_acts.detach().float())
         reg = vocab_manifold_reg(model, thoughts, vocab_reg, vocab_reg_entropy_floor)
 
         # Optimize `.loss` (the trainable term); `.ce_loss` is detached (logging only).
